requirements.txt
numpy==1.16.4
stripe==2.37.2
Flask==1.1.1
Keras==2.2.5
tensorflow==1.13.1
Werkzeug==0.15.4
gunicorn==19.9.0
Pillow==6.2.1
